adjust nuget.config paths
