+++ src/Microsoft.DotNet.TestFramework/TestAssetInfo.cs
@@ -6,13 +6,6 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Runtime.CompilerServices;
-using System.Threading;
-using System.Threading.Tasks;
-using Microsoft.DotNet.Cli.Utils;
-using Microsoft.DotNet.PlatformAbstractions;
-using NuGet.Common;
-using PathUtility = Microsoft.DotNet.Tools.Common.PathUtility;
-using System.Xml.Linq;
 
 namespace Microsoft.DotNet.TestFramework
 {
@@ -57,8 +50,6 @@ namespace Microsoft.DotNet.TestFramework
             DotnetExeFile = dotnetExeFile;
 
             ProjectFilePattern = projectFilePattern;
-
-            //throw new Exception($"root = {_root}\nassetName = {_assetName}\ndotnetExeFile = {_dotnetExeFile}\nprojectFilePattern = {_projectFilePattern}\ndataDir = {_dataDirectory}\ndirectoriesToExclude = {string.Join<DirectoryInfo>(";",_directoriesToExclude)}");
         }
 
         public TestAssetInstance CreateInstance([CallerMemberName] string callingMethod = "", string identifier = "")
@@ -87,87 +78,6 @@ namespace Microsoft.DotNet.TestFramework
 #endif
             return new DirectoryInfo(Path.Combine(baseDirectory, callingMethod + identifier, AssetName));
         }
-
-        private static string RebasePath(string path, string oldBaseDirectory, string newBaseDirectory)
-        {
-            path = Path.IsPathRooted(path) ? PathUtility.GetRelativePath(PathUtility.EnsureTrailingSlash(oldBaseDirectory), path) : path;
-            return Path.Combine(newBaseDirectory, path);
-        }
-
-        private bool IsAncestor(FileInfo file, DirectoryInfo maybeAncestor)
-        {
-            var dir = file.Directory;
-            do
-            {
-                if (string.Equals(maybeAncestor.FullName, dir.FullName, StringComparison.OrdinalIgnoreCase))
-                {
-                    return true;
-                }
-
-                dir = dir.Parent;
-            }
-            while (dir != null);
-
-            return false;
-        }
-
-        //private void DoCopyFiles()
-        //{
-        //    Console.WriteLine($"TestAsset CopyFiles '{AssetName}'");
-
-        //    _operationDirectory.Refresh();
-        //    if (!_operationDirectory.Exists)
-        //    {
-        //        _operationDirectory.Create();
-        //    }
-        //    else
-        //    {
-        //        if (_operationDirectory.GetFiles().Any())
-        //        {
-        //            throw new Exception("operation files folder not empty");
-        //        }
-        //    }
-
-        //    foreach (var f in GetOriginalFileList())
-        //    {
-        //        string destinationPath = RebasePath(f.FullName, Root.FullName, _operationDirectory.FullName);
-        //        var destinationDir = new FileInfo(destinationPath).Directory;
-        //        if (!destinationDir.Exists)
-        //        {
-        //            destinationDir.Create();
-        //        }
-        //        if (string.Equals(f.Name, "nuget.config", StringComparison.OrdinalIgnoreCase))
-        //        {
-        //            var doc = XDocument.Load(f.FullName, LoadOptions.PreserveWhitespace);
-        //            foreach (var v in doc.Root.Element("packageSources").Elements("add").Attributes("value"))
-        //            {
-        //                if (!Path.IsPathRooted(v.Value))
-        //                {
-        //                    string fullPath = Path.GetFullPath(Path.Combine(f.Directory.FullName, v.Value));
-        //                    if (!IsAncestor(new FileInfo(fullPath), Root))
-        //                    {
-        //                        v.Value = fullPath;
-        //                    }
-        //                }
-                        
-        //                //throw new Exception($"\nvalue = {v.Value}\n" +
-        //                //    $"f.dir = {f.Directory.FullName}\n" +
-        //                //    $"fullPath = {fullPath}");
-                        
-        //            }
-
-        //            using (var file = new FileStream(destinationPath, FileMode.CreateNew, FileAccess.ReadWrite))
-        //            {
-        //                doc.Save(file, SaveOptions.None);
-        //            }
-        //        }
-        //        else
-        //        {
-        //            f.CopyTo(destinationPath);
-        //        }
-        //    }
-        //}
-
 
         private void ThrowIfTestAssetDoesNotExist()
         {
